bindings: Correct confusing typo in buffer checks
The error message in the following (incorrect) call was confusing:

    const sodium = require('sodium-native'); void 0;
    const public_key = Buffer.alloc( sodium.crypto_sign_PUBLICKEYBYTES );
    const secret_key = Buffer.alloc( sodium.crypto_sign_PUBLICKEYBYTES );
    const seed = crypto.randomBytes( sodium.crypto_sign_PUBLICKEYBYTES );
    sodium.crypto_sign_seed_keypair( public_key, secret_key, seed );

    Error: secret_key must be a buffer of size crypto_sign_PUBLICKEYBYTES

This commit fixes the error message to indicate the correct named
length. The above is the only example of really misleading output:
crypto_sign_SECRETKEYBYTES is 64, but crypto_sign_PUBLICKEYBYTES is 32.

A few other similar typos are fixed, but crypto_sign_SEEDBYTES is also 32,
so other than being inconsistent, it was not too confusing.